+++ pharma/src/main/java/com/example/pharma/dto/auth/login/LoginSession.java
@@ -6,5 +6,5 @@ public record LoginSession(
         String email,
         UserRole role,
         String codeHash,
-        int attempts
+        Long attempts
 ) {}
