+++ pharma/src/main/java/com/example/pharma/repository/Catalog/BrandRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer> { }
+public interface BrandRepository extends JpaRepository<Brand, Long> { }
