+++ pharma/src/main/java/com/example/pharma/repository/Core/CustomerProfileRepository.java
@@ -3,4 +3,4 @@ package com.example.pharma.repository.Core;
 import com.example.pharma.model.entity.core.CustomerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerProfileRepository extends JpaRepository<CustomerProfile, Integer> {}
+public interface CustomerProfileRepository extends JpaRepository<CustomerProfile, Long> {}
