+++ pharma/src/main/java/com/example/pharma/repository/Core/PharmacistProfileRepository.java
@@ -3,4 +3,4 @@ package com.example.pharma.repository.Core;
 import com.example.pharma.model.entity.core.PharmacistProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PharmacistProfileRepository extends JpaRepository<PharmacistProfile, Integer> {}
+public interface PharmacistProfileRepository extends JpaRepository<PharmacistProfile, Long> {}
