+++ pharma/src/main/java/com/example/pharma/repository/Notification/MedicationReminderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MedicationReminderRepository extends JpaRepository<MedicationReminder, Integer> {
-    List<MedicationReminder> findByCustomer_UserId(Integer userId);
+public interface MedicationReminderRepository extends JpaRepository<MedicationReminder, Long> {
+    List<MedicationReminder> findByCustomer_UserId(Long userId);
 }
