+++ pharma/src/main/java/com/example/pharma/repository/P2P/P2PListingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface P2PListingRepository extends JpaRepository<P2PListing, Integer> {
-    List<P2PListing> findBySeller_UserId(Integer userId);
+public interface P2PListingRepository extends JpaRepository<P2PListing, Long> {
+    List<P2PListing> findBySeller_UserId(Long userId);
 }
