+++ pharma/src/main/java/com/example/pharma/repository/Pharmacy/PharmacyAddressRepository.java
@@ -3,4 +3,4 @@ package com.example.pharma.repository.Pharmacy;
 import com.example.pharma.model.entity.pharmacy.PharmacyAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PharmacyAddressRepository extends JpaRepository<PharmacyAddress, Integer> {}
+public interface PharmacyAddressRepository extends JpaRepository<PharmacyAddress, Long> {}
